refactor: modularize server middleware into separate files for better maintainability
